Move lobby classes to lobby folder.
