+++ src/main/java/client/command/Join.java
@@ -1,0 +1,4 @@
+package client.command;
+
+public class Join {
+}
